+++ components/documentUpload/uploadDocument/ProgressBar.js
